Cambio a tipos en casting

+++ src/main/java/modules/gestionStock/ejb/InsumoEJB.java
@@ -51,11 +51,11 @@ public class InsumoEJB {
     }
 
     private long buscarUltimoID() {
-        TypedQuery<Integer> q = (TypedQuery<Integer>) entityManager.createQuery("SELECT MAX(i.idInsumo) FROM Insumo i");
+        TypedQuery<Long> q = (TypedQuery<Long>) entityManager.createQuery("SELECT MAX(i.idInsumo) FROM Insumo i");
         if(q.getSingleResult() == null){
             return 0;
         }
-        return (Integer) q.getSingleResult();
+        return q.getSingleResult();
     }
